Added faucet controller
