+++ src/http/controllers/index.ts
@@ -2,6 +2,7 @@ export * from './accounts.controller';
 export * from './beams.controller';
 export * from './blocks.controller';
 export * from './core.controller';
+export * from './faucet.controller';
 export * from './governance.controller';
 export * from './health.controller';
 export * from './search.controller';
